Combines Integer and Float monkey-patches into Numeric

+++ lib/niceties/numeric.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-# Niceties for the Integer class
-class Integer
+# Niceties fot the Numeric class
+class Numeric
   def percent_of(other)
     (other / 100.0) * to_f
   end
